Solver evolution toward double instead of float

+++ common.moplaf.macroplanner/com.misc.common.moplaf.macroplanner.solver.emf.edit/src/com/misc/common/moplaf/macroplanner/solver/provider/LPProductItemProvider.java
@@ -206,7 +206,7 @@ public class LPProductItemProvider extends LPTimeLineItemProvider implements IIt
 
 		@Override
 		public float getEventAmount(LPProductBucket bucket, int moment) {
-			return bucket.getConsumed().getSelectedSolutionValue();
+			return (float) bucket.getConsumed().getSelectedSolutionValue();
 		}
 
 		@Override
